components_repo/dspy-prompt-optimizer/requirements.txt
dspy
google-generativeai>=0.3.0
pandas>=2.0.0
python-dotenv>=1.0.0
